requirements.txt
APScheduler==3.6.3
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
cachetools==4.2.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
fake-useragent==0.1.11
greenlet==1.1.2
h11==0.13.0
idna==3.3
lxml==4.9.1
numpy==1.23.1
outcome==1.2.0
pandas==1.4.3
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-telegram-bot==13.13
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
scrapper-boilerplate==0.2.8
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.39
tornado==6.2
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.10
webdriver-manager==3.8.2
wsproto==1.1.0
